requirements.txt
dash==2.9.2
numpy==1.19.5
Pillow==9.5.0
tensorflow==2.4.0
tensorflow_intel==2.11.0
